Combine system.ts into effect.ts
