+++ packages/reactivity/src/dep.ts
@@ -1,7 +1,7 @@
 import { isArray, isIntegerKey, isMap, isSymbol } from '@vue/shared'
 import { type TrackOpTypes, TriggerOpTypes } from './constants'
 import { onTrack, triggerEventInfos } from './debug'
-import { Dependency, type Link, System, endBatch, startBatch } from './system'
+import { Dependency, type Link, System, endBatch, startBatch } from './effect'
 
 class Dep implements Dependency {
   _subs: Link | undefined = undefined
